A3 Modifications (Assignment Editor)
Since we integrated a sample database, instead of having just a filler "everyone" option, I made the component dynamically render the students in the JSON we created. It provides everyone as an option by default (greying out the individual students), and if everyone is unchecked, all students can individually be selected. Individual student selections will be displayed in the component. It is all visual, and nothing is saved to the db.